Move I_GetTime() to IWRAM
