+++ include/d_main.h
@@ -59,7 +59,6 @@ void D_DoAdvanceDemo (void);
 // BASE LEVEL
 //
 
-void D_Display(void);
 void D_PageTicker(void);
 void D_StartTitle(void);
 void D_DoomMain(void);
