+++ source/i_system.c
@@ -51,50 +51,6 @@
 #include "global_data.h"
 
 
-/* CPhipps - believe it or not, it is possible with consecutive calls to 
- * gettimeofday to receive times out of order, e.g you query the time twice and
- * the second time is earlier than the first. Cheap'n'cheerful fix here.
- * NOTE: only occurs with bad kernel drivers loaded, e.g. pc speaker drv
- */
-int I_GetTime(void)
-{
-    int thistimereply;
-
-#ifndef __arm__
-    struct timeval tv;
-    struct timezone tz;
-
-    gettimeofday(&tv, &tz);
-
-    thistimereply = (tv.tv_sec * TICRATE + (tv.tv_usec * TICRATE) / 1000000);
-
-    thistimereply = (thistimereply & 0xffff);
-#else
-    thistimereply = I_GetTime_e32();
-#endif
-
-    if (thistimereply < _g->lasttimereply)
-    {
-        _g->basetime -= 0xffff;
-    }
-
-    _g->lasttimereply = thistimereply;
-
-
-    /* Fix for time problem */
-    if (!_g->basetime)
-    {
-        _g->basetime = thistimereply;
-        thistimereply = 0;
-    }
-    else
-    {
-        thistimereply -= _g->basetime;
-    }
-
-    return thistimereply;
-}
-
 /* cphipps - I_GetVersionString
  * Returns a version string in the given buffer 
  */
