+++ source/i_system_gba.cpp
@@ -316,14 +316,7 @@ int I_GetVideoHeight_e32()
     return 160;
 }
 
-//**************************************************************************************
 
-int I_GetTime_e32(void)
-{
-    int thistimereply = REG_TM3CNT_L;
-
-    return thistimereply;
-}
 
 //**************************************************************************************
